CC 107: Completely changed the data structure and much of how it works.
Instead of a simple 2D array of numbers, it now maintains a grid of
linked objects (kind of like a 2D double-linked list), which can start
small and expand as necessary to keep all the sand piles.

The sand is no longer added all at once at the start, instead we add a
smaller amount of sand (adjustable by slider) on each frame. It also
finishes a complete toppling per frame, so that every frame displayed
is actually a valid sandpile group (only 0-3), instead of showing the
intermediate states. (Though I'm considering changing that again.)

It no longer topples by doing complete passes over every pile, but
instead maintains a queue of which piles to topple, based on which ones
we have added enough sand to. It also does not copy the board for each
toppling, as (I believe) the math works out to not really need it - as
long as we don't overwrite any piles, only adjust them (add/subtract).

This is generally quite a bit faster, as it usually doesn't need to
topple all that many piles. However, it makes the framerate uneven,
since sometimes there _are_ a lot of piles to topple before it's done.